layer config refactor
