+++ elements/layercontrol/src/helpers/update-url.js
@@ -1,0 +1,22 @@
+/**
+ * Update layer's URL using input values
+ * @param {String} url
+ * @param {Object} values
+ */
+export default function updateUrl(url, values) {
+  const searchParams = new URL(url).searchParams;
+
+  Object.entries(values).forEach(([key, value]) => {
+    if (typeof value === "object" && !Array.isArray(value) && value !== null) {
+      Object.keys(value).forEach((k) => {
+        searchParams.set(k, value[k]);
+      });
+    } else searchParams.set(key, value);
+  });
+
+  const urlWithPath = url.split("?")[0];
+  const searchParamsStr = searchParams.toString();
+  const newUrl = `${urlWithPath}?${searchParamsStr}`;
+
+  return newUrl;
+}
