+++ elements/layercontrol/src/methods/layer-config/index.js
@@ -1,0 +1,1 @@
+export { default as dataChangeMethod } from "./data-change";
